develop commit 4

+++ src/blogs/blogs.service.ts
@@ -1,7 +1,7 @@
 import { RefType, SortOrder } from "mongoose";
 import { BlogModel } from "./schema/blog.schema";
-import { IBlog } from "./interface/blog.interface";
-import { Inject, Injectable } from "@nestjs/common";
+import { IBlog, IBlogWithUserId } from "./interface/blog.interface";
+import { ForbiddenException, Inject, Injectable } from "@nestjs/common";
 import { BlogsRepository } from "./blogs.repository";
 import { CreateBlogDto } from "./dto/create-blog.dto";
 import { UpdateBlogDto } from "./dto/update-blog.dto";
@@ -32,29 +32,46 @@ export class BlogsService {
         return this.blogRepository.findAll(searchNameTerm, skip, pageSize, sortBy, sortDirection, userId);
     }
 
-    public async findOne(id: RefType): Promise<IBlog | undefined> {
+    public async findOne(id: RefType | string): Promise<IBlog | undefined> {
         const blog = await this.blogRepository.find(id);
         if (!blog) throw new Error();
 
         return blog;
     }
 
-    public async update(id: RefType, updateBlogDto: UpdateBlogDto): Promise<IBlog | undefined> {
-        const updateBlog: IBlog | undefined | null = await this.blogRepository.updateBlog(id, updateBlogDto);
-        if (updateBlog) return updateBlog;
-        throw new Error();
+    public async update(id: RefType, userId: string, updateBlogDto: UpdateBlogDto): Promise<IBlog | undefined> {
+        const blog = (await this.blogRepository.find(id)) as IBlogWithUserId;
+        if (!blog) throw new Error();
+        if (blog.userId === userId) {
+            return await this.blogRepository.updateBlog(id, updateBlogDto);
+        }
+        throw new ForbiddenException();
     }
 
-    public async delete(id: RefType): Promise<IBlog> {
-        const deleteBlog = await this.blogRepository.delete(id);
-        if (deleteBlog) return deleteBlog;
-        throw new Error();
+    public async delete(id: RefType, userId: string): Promise<IBlog> {
+        const blog = (await this.blogRepository.find(id)) as IBlogWithUserId;
+        if (!blog) throw new Error();
+        if (blog.userId === userId) {
+            return await this.blogRepository.delete(id);
+        }
+        throw new ForbiddenException();
     }
 
     public async getTotalCountForBlogs(searchNameTerm: string | undefined | object, userId: string): Promise<number> {
         if (searchNameTerm) searchNameTerm = { name: { $regex: new RegExp(`.*${searchNameTerm}.*`, "i") } };
 
         return await this.blogRepository.getBlogsCount(searchNameTerm, userId);
+    }
+
+    public async bindingBlogWithUser(id: string, userId: string) {
+        const blogWithoutOwner = (await this.findOne(id)) as IBlogWithUserId;
+        if (!blogWithoutOwner) {
+            throw new Error();
+        }
+        if (blogWithoutOwner.userId === null) {
+            return await this.blogRepository.updateBlogByBind(id, userId);
+        }
+        throw new Error();
     }
 
     public async testingDelete(): Promise<void> {
